ldappool: force-connect server during _bind()
In case there is no StartTLS and no bind DN, _bind() is effectively a
no-op.

Since python-ldap connections are lazy this means that connections returned
by ConnectionManager are not necessary 'real' connections: it might happen
that no operations have been done and thus no real connection has been
established to the server. This effectively prevents ldappool-based server
fallback from working.

This change adds a whoami_s() call which is effectively a no-op, but
forces connecting to the server.